fix: preserve activity actor scope evidence
