+++ apps/api/src/handlers/workspaces/read-overview-activity-actors.ts
@@ -1,13 +1,9 @@
 import { Result } from "better-result";
-import { and, asc, eq, gt, ilike, or, sql } from "drizzle-orm";
 import { t } from "elysia";
 
-import { member, user } from "@/api/db/auth-schema";
-import { auditLogs } from "@/api/db/schema";
 import { createSafeHandler } from "@/api/lib/api-handlers";
 import type { HandlerConfig } from "@/api/lib/api-handlers";
 import { HandlerError } from "@/api/lib/errors/tagged-errors";
-import { escapeLike } from "@/api/lib/escape-like";
 import { LIMITS } from "@/api/lib/limits";
 import {
   createCursorPage,
@@ -15,7 +11,7 @@ import {
   encodePaginationCursor,
 } from "@/api/lib/pagination";
 
-import { visibleActivityCondition } from "./read-overview-activity.query";
+import { readOverviewActivityActorRows } from "./read-overview-activity.query";
 
 const config = {
   permissions: { workspace: ["read"] },
@@ -31,9 +27,6 @@ const config = {
     search: t.Optional(t.String({ maxLength: 256 })),
   }),
 } satisfies HandlerConfig;
-
-const historicalActorId = () =>
-  sql<string>`coalesce(${auditLogs.performerId}, ${auditLogs.userId})`;
 
 const decodeActorCursor = (cursor: string, search: string): string | null => {
   const parts = decodePaginationCursor(cursor);
@@ -60,61 +53,21 @@ const readOverviewActivityActors = createSafeHandler(
     }
 
     const limit = query.limit ?? LIMITS.matterActivityActorPageSizeDefault;
-    const result = yield* Result.await(
-      safeDb(async (tx) => {
-        const actorId = historicalActorId();
-        const conditions = [
-          eq(auditLogs.organizationId, session.activeOrganizationId),
-          eq(auditLogs.workspaceId, workspaceId),
-          eq(auditLogs.performerType, "user"),
-          visibleActivityCondition(),
-        ];
-        if (afterActorId !== null) {
-          conditions.push(gt(actorId, afterActorId));
-        }
-        const historicalActors = tx
-          .selectDistinct({ id: actorId.as("actor_id") })
-          .from(auditLogs)
-          .where(and(...conditions))
-          .as("historical_activity_actors");
-        const identityCondition =
-          search === ""
-            ? sql`true`
-            : (or(
-                ilike(user.name, `%${escapeLike(search)}%`),
-                ilike(user.email, `%${escapeLike(search)}%`),
-              ) ?? sql`false`);
-
-        const actorRows = await tx
-          .selectDistinct({
-            deletedAt: user.deletedAt,
-            email: user.email,
-            id: historicalActors.id,
-            image: user.image,
-            name: user.name,
-          })
-          .from(historicalActors)
-          .leftJoin(
-            member,
-            and(
-              eq(member.userId, historicalActors.id),
-              eq(member.organizationId, session.activeOrganizationId),
-            ),
-          )
-          // The actor ID comes from an organization-and-workspace-scoped audit
-          // row, so attribution remains authorized after membership ends.
-          .leftJoin(user, eq(user.id, historicalActors.id))
-          .where(identityCondition)
-          .orderBy(asc(historicalActors.id))
-          .limit(limit + 1);
-        const page = createCursorPage({
-          rows: actorRows,
-          limit,
-          cursorForItem: ({ id }) => encodePaginationCursor([search, id]),
-        });
-        return page;
+    const actorRows = yield* Result.await(
+      readOverviewActivityActorRows({
+        afterActorId,
+        limit,
+        organizationId: session.activeOrganizationId,
+        safeDb,
+        search,
+        workspaceId,
       }),
     );
+    const result = createCursorPage({
+      rows: actorRows,
+      limit,
+      cursorForItem: ({ id }) => encodePaginationCursor([search, id]),
+    });
 
     return Result.ok({
       items: result.items.map(({ deletedAt, email, id, image, name }) => ({
