+++ scripts/ratchet-baseline.json
@@ -116,10 +116,11 @@
     "files": {}
   },
   "security-guards-no-unscoped-user-query-suppressions": {
-    "count": 2,
+    "count": 3,
     "files": {
       "apps/api/src/handlers/operator/query.ts": 1,
-      "apps/api/src/handlers/workspaces/list.ts": 1
+      "apps/api/src/handlers/workspaces/list.ts": 1,
+      "apps/api/src/handlers/workspaces/read-overview-activity.query.ts": 1
     }
   },
   "security-guards-require-secure-document-response-suppressions": {
