+++ scripts/suppression-waivers.json
@@ -109,6 +109,16 @@
       "count": 1,
       "reason": "joined via RLS-scoped workspaceMembers filtered by org-derived wsIds (see comment above)",
       "evidence": "apps/api/src/handlers/workspaces/list.integration.test.ts"
+    },
+    {
+      "id": "SW-0012",
+      "kind": "permanent",
+      "rule": "security-guards/no-unscoped-user-query",
+      "file": "apps/api/src/handlers/workspaces/read-overview-activity.query.ts",
+      "symbol": "import @/api/db/auth-schema",
+      "count": 1,
+      "reason": "actor IDs come only from audit rows already scoped to the authorized organization and workspace; membership joins would erase retained attribution after membership ends.",
+      "evidence": "apps/api/src/handlers/workspaces/read-overview-activity.integration.test.ts"
     }
   ]
 }
